move service back to main package
