+++ src/test/java/uk/gov/hmcts/reform/em/hrs/service/impl/AuditEntryServiceTests.java
@@ -20,6 +20,7 @@ import uk.gov.hmcts.reform.em.hrs.domain.HearingRecordingShareeAuditEntry;
 import uk.gov.hmcts.reform.em.hrs.repository.HearingRecordingAuditEntryRepository;
 import uk.gov.hmcts.reform.em.hrs.repository.HearingRecordingSegmentAuditEntryRepository;
 import uk.gov.hmcts.reform.em.hrs.repository.ShareesAuditEntryRepository;
+import uk.gov.hmcts.reform.em.hrs.service.AuditEntryService;
 import uk.gov.hmcts.reform.em.hrs.service.SecurityService;
 
 import java.util.List;
