+++ src/test/java/uk/gov/hmcts/reform/em/hrs/service/impl/BlobStorageDeleteServiceTest.java
@@ -16,6 +16,7 @@ import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.testcontainers.shaded.org.apache.commons.lang3.RandomStringUtils;
 import uk.gov.hmcts.reform.em.hrs.dto.HearingSource;
+import uk.gov.hmcts.reform.em.hrs.service.BlobStorageDeleteService;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
