+++ src/test/java/uk/gov/hmcts/reform/em/hrs/service/impl/HearingRecordingServiceImplTest.java
@@ -11,6 +11,7 @@ import uk.gov.hmcts.reform.em.hrs.domain.HearingRecording;
 import uk.gov.hmcts.reform.em.hrs.domain.HearingRecordingSegment;
 import uk.gov.hmcts.reform.em.hrs.dto.HearingSource;
 import uk.gov.hmcts.reform.em.hrs.repository.HearingRecordingRepository;
+import uk.gov.hmcts.reform.em.hrs.service.BlobStorageDeleteService;
 
 import java.util.ArrayList;
 import java.util.Collections;
