+++ src/test/java/uk/gov/hmcts/reform/em/hrs/service/impl/ShareAndNotifyServiceImplTest.java
@@ -8,6 +8,7 @@ import uk.gov.hmcts.reform.em.hrs.exception.ValidationErrorException;
 import uk.gov.hmcts.reform.em.hrs.model.CaseDocument;
 import uk.gov.hmcts.reform.em.hrs.model.CaseHearingRecording;
 import uk.gov.hmcts.reform.em.hrs.repository.HearingRecordingRepository;
+import uk.gov.hmcts.reform.em.hrs.service.AuditEntryService;
 import uk.gov.hmcts.reform.em.hrs.service.NotificationService;
 import uk.gov.hmcts.reform.em.hrs.service.ShareeService;
 import uk.gov.hmcts.reform.em.hrs.service.ccd.CaseDataContentCreator;
